General: Tool panel clean-ups and split geometry alignment linking into smaller components
