+++ ui/src/tool-panel/location-track/location-track-infobox-duplicate-of.tsx
@@ -6,34 +6,29 @@ import styles from './location-track-infobox.scss';
 export type LocationTrackInfoboxDuplicateOfProps = {
     existingDuplicate: LayoutLocationTrack | undefined;
     duplicatesOfLocationTrack: LayoutLocationTrackDuplicate[] | undefined;
-}
+};
 
 export const LocationTrackInfoboxDuplicateOf: React.FC<LocationTrackInfoboxDuplicateOfProps> = ({
     existingDuplicate,
     duplicatesOfLocationTrack,
 }: LocationTrackInfoboxDuplicateOfProps) => {
-    return (
-        <React.Fragment>
-            {
-                existingDuplicate ?
+    return existingDuplicate ? (
+        <LocationTrackLink
+            locationTrackId={existingDuplicate.id}
+            locationTrackName={existingDuplicate.name}
+        />
+    ) : duplicatesOfLocationTrack ? (
+        <ul className={styles['location-track-infobox-duplicate-of__ul']}>
+            {duplicatesOfLocationTrack.map((duplicate) => (
+                <li key={duplicate.id}>
                     <LocationTrackLink
-                        locationTrackId={existingDuplicate.id}
-                        locationTrackName={existingDuplicate.name}
+                        locationTrackId={duplicate.id}
+                        locationTrackName={duplicate.name}
                     />
-                    : duplicatesOfLocationTrack ?
-                       <ul className={styles['location-track-infobox-duplicate-of__ul']}> {
-                            duplicatesOfLocationTrack.map(duplicate =>
-                               <li key={duplicate.id}>
-                                   <LocationTrackLink
-                                       locationTrackId={duplicate.id}
-                                       locationTrackName={duplicate.name}
-                                   />
-                               </li>
-                            )
-                        }
-                       </ul>
-                        : ''
-            }
-        </React.Fragment>
+                </li>
+            ))}
+        </ul>
+    ) : (
+        <React.Fragment />
     );
 };
